CHANGE: deprecated logical_tie_masks property. Use SilenceMask instead.
The reason for the change has to do with considerably increased functionality.

Example:

OLD:

    rmakers.TaleaRhythmMaker(
        beam_specifier=rmakers.BeamSpecifier(
            beam_each_division=True,
        ),
        logical_tie_masks=[rmakers.SilenceMask([2], period=3)]
    )

NEW:

    pattern = abjad.Pattern([2], period=3)
    rmakers.TaleaRhythmMaker(
        rmakers.SilenceMask(
            selector=abjad.select().logical_ties()[pattern]
        ),
        beam_specifier=rmakers.BeamSpecifier(
            beam_each_division=True,
        ),
    )

COMMENT. Initialize silence mask and sustain mask objects as shown above. Pass
an arbitrary number of masks in to rhythm-maker __init__() method.
Two things are then important during interpretation:

    1. Rhythm-makers interpret masks IN ORDER.

    2. Mask selector can be set to anything.

This means that arbitrarily complex patterns of masking are now possible, for
the first time.